Applied patch from JIRA issue Survey Implementation for testing scenarios of ecommerce application (OFBIZ-2100).
Initial work is done by Harsha Chadhar (data preperation, service, form, screens) with the help of Pranay Pandey.Thanks to both & special thanks to Anil Patel for providing useful details in b/w Implementation. Then I took the charge as there were some issues here and there.
(Two points are covered in this patch and remaining will be covered soon. So be patient for the complete availability of the feature :-) ) 

After completion of this work we should provide the Document on confluence to use this feature.

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk/specialpurpose/projectmgr@767400 13f79535-47bb-0310-9956-ffa450edef68